pyproject.toml
[project]
name = "employement"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sanic (>=25.3.0,<26.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "tortoise-orm (>=0.25.1,<0.26.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "jwt (>=1.4.0,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
